refactor(processor): extract cache payload validation
Move cached image and attachment usability checks into a processor helper so processDocument can focus on cache flow instead of filesystem validation details.
